requirements.txt
streamlit==1.44.1
pandas==2.2.3
plotly==5.18.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pymongo==4.6.1
tabulate==0.8.10
altair==5.5.0
